refactor: move contract errors to errors folder
